ub:ubus: Fixing the bug of concurrent hotplug and device reset
driver inclusion
category: bugfix
bugzilla: https://atomgit.com/openeuler/kernel/issues/6695
CVE: NA

-----------------------------------------------------------

In hotplug and linkup events, the same device is enumerated only once.

Fixes: 8f4f082b147e ("ub:ubus: Support hotplug service driver")

+++ drivers/ub/ubus/services/hotplug/hotplug_core.c
@@ -637,7 +637,7 @@ static void ubhp_handle_present(struct ub_slot *slot)
 		return;
 	}
 
-	if (ubhp_handle_link_up(slot))
+	if (!slot->r_uent && ubhp_handle_link_up(slot))
 		goto poweroff;
 
 	ubhp_set_indicators(slot, INDICATOR_ON, INDICATOR_NOOP);
